+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Export journal entry categories name
+
 [13.1.0] - 2025-10-08
 
 ### Added